pyproject.toml
[project]
name = "nanoSNN"
version = "0.1.0"
description = "Vision SNN training framework inspired by nanoLMengine"
readme = "README.md"
requires-python = ">=3.10,<3.13"
dependencies = [
    "torch==2.6.0",
    "torchvision==0.21.0",
    "torchaudio==2.6.0",
    "lightning>=2.0",
    "timm",
    "spikingjelly==0.0.0.0.14",
    "numpy",
    "PyYAML",
    "tensorboard",
    "wandb",
]

[[tool.uv.index]]
name = "pytorch-cu124"
url = "https://download.pytorch.org/whl/cu124"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch-cu124" }
torchvision = { index = "pytorch-cu124" }
torchaudio = { index = "pytorch-cu124" }
